+++ packages/monaco-editor-wrapper/test/wrapper.test.ts
@@ -24,7 +24,6 @@ describe('Test MonacoEditorLanguageClientWrapper', () => {
 
         const app = wrapper.getMonacoEditorApp() as EditorAppClassic;
         expect(app).toBeDefined();
-        expect(app.getAppType()).toBe('classic');
 
         const appConfig = app.getConfig();
         expect(appConfig.automaticLayout).toBeTruthy();